PCHR-3472: Add custom make file to create dev sites
The custom file creates a working copy when download from the civihr repository.
Besides, instead of download the latest CiviHR stable version, it downloads from
staging.

Using this custom file require some special params to be passed to drush, so a
build-dev.sh script was added to help with that