PCHR-3681: Fixing permission issues when installing using drush site-install
During the installation, Drupal will try to set secure default permissions
to the settings file and folder. When installing via the UI, this is often
not possible, as, usually, the webserver user will not be the owner of the
settings file and folder and so it will not be possible to change their
permission.

However, when installing via drush site-install, the user who executes
the command is often the owner of the files, meaning that the secure
permissions will be set. As a result, we might have permission issues
when trying to create the CiviCRM settings file or updating the Drupal
settings file, when finishing the installation. To void that, we now
make sure we have write permissions first and revert them to non-writeable
after finishing any updates.